Add USB PID for Logicbone DFU Bootloader
This adds a USB PID for the DFU bootloader on the Logicbone FPGA
development board. The hardware for this project is licensed under
the CERN OHL v1.2 license and the hardware design files are hosted
at https://github.com/oskirby/logicbone

The bootloader project is a fork of the TinyFPGA bootloader,
licesened under the Apache 2.0 license. The verilog sources are
hosted at https://github.com/oskirby/tinydfu-bootloader