USB IDs for the Osmocom E1 tracer
This was originally a "one-off" at CCC, but is getting a dedicated board
with everything integrated on it.

The wiki page is just a stub for now, but still has links to the published
hardware (eagle) along with the original prototype version page that has
the published fpga gateware / firmware and host software.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>