usb_ids: Request iCEKeeb keyboard firmware IDs
This is a FPGA based keyboard firmware / gateware.

The firmware and gateware is at https://github.com/esden/icekeeb

* Gateware is mostly under "CERN Open Hardware license v2 Permissive",
but some of the cores are under other permissive licenses. Each file
has the license specified in its header.

* Firmware is mostly under LGPL3+, against some files imported from
other places are under more permissive licenses. Each file has the license
specified in its header.

* Some other files/cores imported through submodules might differ from
the above but they are not used / don't end up in the final build, that's
just the nature of submodules.

The hardware is based off the icebreaker-bitsy modules
( https://github.com/icebreaker-fpga/icebreaker/tree/master/hardware/bitsy-v1.1b )
integrated on a custom baseboard (which is not released, at least yet).

But having the firmware/gateware under appropriate license is sufficient
according to conditions listed and the same fw/gw can be used on different
hardware.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>