usb_ids: Request for Nitro FPGA μACM IDs
This is a "drop-in" USB CDC ACM module meant for people that don't want
to dig into USB and just want quick I/O through emulated serial. The plan
is to ship pre-synthesized version of it that can be used by other projects
as a "blackbox" component just like if your FPGA had a ftdi attached to
it.

Project is hosted at :
 https://github.com/no2fpga/no2muacm

License is mostly CERN-OHL-P-2.0 (gateware) / MIT (firmware)
all details are in the repo and in each individual file header
(at least for the submodules I control, I can't dictate how other
people organize their stuff ...)

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>