usb_ids: Requests for IDs for the AMR modem interface using iCE40
The gateware and firmware are in the linked repo and under various
licenses. Generally the gateware modules are under one of the
CERN-OHL-V2 license, and the software under GPL/LGPL/MIT.
Each file has associated SPDX header and full license text for
the various license is in top level doc/.

The associate hardware is also technicaly open hardware but is
not under my control, this targets the icebreaker dev board
manually wired to a AMR slot breakout board, there is no
dedicated hw yet.

At some point it might move repo with dedicated hardware, host
software, ... but for now this is it.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>