request: Add Jumperless to PID list
Jumperless - a jumperless breadboard

Licenses:
Hardware - CERN-OHL-W-2.0
Software - MIT
Documentation - CC-BY-4.0

OSHWA ID: US002417

https://github.com/Architeuthis-Flux/Jumperless

I might need one more PID as the Jumperless shows up as 2 USB endpoints (the other acts as a routable USB-UART bridge), and it might happen that we'll want that to show up as a completely separate device.

I hope it's okay to pick an arbitrary PID to match the current made-up PID I've been using. If not, it's fine and I can make it sequential.