USB PID Request for Baochip series of products
Project Information:

    Repository: https://github.com/baochip/baochip-1x, https://github.com/baochip/dabao, https://github.com/betrusted-io/xous-core/
    License: Apache 2.0 (software), CERN OHL-W-2.0 (hardware)
    Description: Baochip - an open source SoC plus products based on the Baochip-1x

Requested PID Allocation:

  These have been offset from the existing PRs to make merging easier, presuming they will be granted:

  0x1d50 | 0x6196 | [https://github.com/baochip/baochip-1x Baochip-1x SoC bootloader]
  0x1d50 | 0x6197 | [https://github.com/baochip/dabao Dabao breakout board for Baochip-1x]
  0x1d50 | 0x6198 | [https://github.com/betrusted-io/xous-core/ Baosec security token]

I am requesting three total PIDs: one for a bootloader that is common to all products,
plus one for a breakout board, and one for a security token product. The breakout board
and security token both use the same chip, but present to the user as quite different
sets of endpoints, hence the request for two separate PIDs.

Thanks for your consideration!